pyproject.toml
[tool.poetry]
name = "cybergraph"
version = "0.1.0"
description = "Graph-based RAG System for cybersec analysis"
authors = ["Ananth Prathap <22am014@sctce.ac.in>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.0.325"
langgraph = "^0.0.10"
chromadb = "^0.4.15"
openai = "^1.3.0"
networkx = "^3.1"
python-dotenv = "^1.0.0"
pydantic = "^2.4.2"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
python-multipart = "^0.0.6"
pillow = "^10.1.0"
aiohttp = "^3.9.1"
aiofiles = "^23.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
